Improve the vsphere-iso boot_command
In production vsphere hosts with highly variable loads it is difficult to accurately predict how long to wait before sending the boot_command.  By holding down "escape" at the beginning of the boot command you force the VM bios to show the "boot device" menu. From there you can more easily predict when to send the <esc> key that loads the ubuntu graphical boot menu and in-turn be able to predict what keys to press.